+++ README.md
@@ -1,1 +1,43 @@
-# 11_sprint_tmv
+<h1># 11_sprint_tmv</h1>
+<h2>Изучение основ Webpack и модульности JS</h2>
+<h3>version 0.0.2</h3>
+
+<p> Изучение принципов работы со сборщиком проектов на основе Webpack. Процесс работы над проектом состаял из нескольких эатпов:
+1. Создание репозитория на Github.
+2. Перенос репозитория в локальную папку и копирование файлов для работы.
+3. Установка и настройка WebPack при помощи командной строки
+    3.1. Установка пакетов
+    3.2. Описание логики работы установленных пакетов.
+4. Создание дополнительных файлов для корректоной работы:
+    4.1. .nojekyll - Для корректной работы сборщика с файлами по методологии БЭМ
+    4.2 .gitignore - Для исключения выгрузки "лишних" файлов и папок в рабочую ветку.
+5. Разбиение кода JS файла на модули и их послудующая связка по средствам import и export.
+6. Выгрузка готового кода на github pages.
+
+Для разработки проекта использовалось 3 ветки:
+1. master - Главная ветка, на данный момент туда выгружена изначальная вариация проекта, после прохождения ревью, ветка будет смерджина с develop.
+2. develop - ветка, в которой производилась сама работа
+3. gh-pages - специальная ветка для размещения рабочего варианта работы на Github
+
+В процессе работы использоваись 3 сборки:
+1. build - сборка для продакшен версии, собирается в специальной папке dist, для сборки, необходимо в командной строке ввести команду npm run build
+2. dev - сборка для разработки (что особо удобно при подключении режима горячей сборки). Сборка открывается на локальном сервере по адресу http://localhost:8080, для выполнения сборки, в командной строке необходимо ввести npm run dev
+3. deploy - сборка для непосредственной публикации на сервере. Для выполнения сборки, в необходимо ввести в консоли команду npm run deploy, после чего сайт выгрузится на сервер.
+
+Планы по доработке:
+1. Работа над модульностью - разбивка файла прошла успешна, но необходимо скорретировать код.
+2. Изучения дополнительных пакетов для WebPack
+
+Готовую работу можно посмотреть по адресу https://Mturin.github.io/11_sprint_tmv/ </p>
+
+<h4>Для ревьюверов</h4>
+<p>Заранее благодарю за выденное на проверку работы время и приношу свои извинения, что направил под самый дедлайн. Очень искренне наеюсь, что работа выполнена корректно и критических замечаний не будет и я выйду на "диплом" без потери месяца.
+P.S.  В любом случае, так как это последний спринт для меня перед дипломом, хочу выразить Вам огромную благодарность за Вашу работу. Я искренне не понимаю откуда у Вас столько терпения и выдержки что бы читать наш корявый ученический код и при этом сохранять самообладание и давать полезные комментарии. Да, с одной стороны это Ваша работа, но с другой, человеческое отношение и проявление заботы не купишь за деньги. Мне было очень приятно читать Ваши комментарии, надеюсь, если все таки разработка станет для меня основной работой, что я не подведу Вас. СПАСИБО ВАМ ОГРОМНОЕ, с уважением Тюренков Максим. </p>
+
+<h5>Ощущения от работы</h5>
+<p> Кошмар.... до последнего момента, думал что не потяну. Казалось что после JS и регулярок не может быть уже ничего хуже, но нет =)))). Проблемы были слудующие:
+1. Webpack - Ооооооочень сухой материал в тренажере. Много информации, множество настроек, а как это все работает и для чего не понятно. Очень не хватало видео "от и до". Спасло то, что в итоге все расписано по шагам, что позволило воспроизвести настройку на своем проекте.
+2. Модульность - разбивка своего JS файла была подобно хождению по минному полю, класс перенес, пропустил переменную все слетело. Конечно в итоге надо это делать сразу, во время написания, к чему и буду стримиться в дальнейшем. 
+3. Скрытые ошибки, которые не так легко было обнаружить, файл с картинкой Карачаевска, а также конкретно в моем коде, была ошибка с одним из файлов стилей попап, но неуверен что это было изначально заложено и при этом не помню чтобы сам добавлял этот файл.
+4. Общим итогом сносил и начинал все заново 4 раза, финальный вариант начал делать в птницу вечером с 17-00, остановился в субботу утром в 6 утра...., в убботу вечером продолжил и пишу этот финальный текст уже в 4 утра воскресенья. =)))))
+5. Устал очень, но дико рад, что все получилось (надеюсь)</p>
