+++ src/blocks/popup/popup.css
@@ -5,7 +5,7 @@
 @import url(./__input/popup__input.css);
 @import url(./__button/popup__button.css);
 @import url(./_is-opened/popup_is-opened.css);
-@import url(./__button/popup__button\ _active.css);
+@import url(./__button/popup__button_active.css);
 
 .popup {
   position: fixed;
