+++ src/index.html
@@ -5,12 +5,13 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
-  <link rel="stylesheet" href="./pages/index.css">
+  <!-- <link rel="stylesheet" href="./pages/index.css"> -->
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
 </head>
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
 
@@ -54,7 +55,7 @@
     </div>
     <div class="popup popup__addPlace">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="new">
             <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Название">
@@ -68,7 +69,7 @@
 
     <div class="popup popup__user">
       <div class="popup__content">
-      <img src="./images/close.svg" alt="" class="popup__close popup__close_user">
+      <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close popup__close_user">
       <h3 class="popup__title">Редактировать профиль</h3>
       <form class="popup__form popup__form_user" name="editUser">
           <input type="text" name="userName" novalidate class="popup__input popup__input_user_name" placeholder="Имя" value='Jaques Causteau'>
@@ -82,7 +83,7 @@
 
     <!-- <div class="popup popup__user">
       <div class="popup__content">
-      <img src="./images/close.svg" alt="" class="popup__close popup__close_avatar">
+      <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close popup__close_avatar">
       <h3 class="popup__title">Аватар</h3>
       <form class="popup__form popup__form_userAvatar" name="avatar">
            <input type="text" name="userAvatar" class="popup__input popup__input_user_avatar" placeholder="ссылка на фото" >
@@ -93,9 +94,10 @@
 
     <div class="popup popup__image">
       <div class="popup__content popup__content_image">
-          <img src="./images/close.svg" alt="" class="popup__close popup__close_image">
+          <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close popup__close_image">
       </div>
     </div>
-  <script src='./script.js'></script>
+    <!-- <script src='./js_module/script.js'></script> -->
+  <script src='<%= htmlWebpackPlugin.files.chunks.main.entry %>'></script>
 </body>
 </html>
