+++ src/pages/index.css
@@ -10,6 +10,7 @@
 @import url(../blocks/place-card/place-card.css);
 @import url(../blocks/popup/popup.css);
 
+
 body {
   background-color: #000;
 }
